Исправления по замечаниям Александра Оруджевева от 05.02.2020
