+++ Readme.md
@@ -4,6 +4,10 @@
 ## Содержание:
 * [Студент](#Студент)
 * [Папка students-testing](#Папка-students-testing)
+* [Папка students-testing-03](#Папка-students-testing-03)
+* [Папка students-testing-shell](#Папка-students-testing-shell)
+* [Папка library](#Папка-library)
+* [Папка library-orm](#Папка-library-orm)
 
 # Студент
 ФИО слушателя: Люлин Вадим Евгеньевич
@@ -11,4 +15,26 @@
 Группа: 2019-11
 
 ## Папка students-testing
+Домашнее задание по теме "Конфигурирование Spring-приложений"
 Программа по проведению тестирования студентов
+
+## Папка students-testing-03
+Домашнее задание по теме "Чёрная магия" Spring Boot
+Перенести приложение для тестирования студентов на Spring Boot
+
+## Папка students-testing-shell
+Домашнее задание
+Перевести приложение для проведения опросов на Spring Shell
+
+## Папка library
+Домашнее задание по теме "DAO на Spring JDBC"
+Приложение хранящее информацию о книгах в библиотеке
+
+## Папка library-orm
+Домашнее задание по теме "Переписать приложение для хранения книг на ORM"
+Приложение хранящее информацию о книгах в библиотеке
+
+## Папка library-nosql
+Домашнее задание по теме "Использовать MongoDB и spring-data для хранения информации о книгах"
+Приложение хранящее информацию о книгах в библиотеке
+
