+++ library-spring-nosql/HELP.md
@@ -1,0 +1,32 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<module type="JAVA_MODULE" version="4" />
+
+## Содержание:
+* [Домашнее задание "Использовать MongoDB и spring-data для хранения информации о книгах"](#Библиотека на Spring Data NoSQL)
+* [Reference Documentation](#Reference Documentation)
+
+# Библиотека на Spring Data NoSQL
+
+Использовать MongoDB и spring-data для хранения информации о книгах
+Цель: Цель: После выполнения ДЗ вы сможете использовать Spring Data MongoDB и саму MongoDB для разработки приложений с хранением данных в нереляционной БД. Результат: Приложение с использованием MongoDB
+Задание может выполняться на основе предыдущего, а может быть выполнено самостоятельно
+
+Требования:
+1. Использовать Spring Data MongoDB репозитории, а если не хватает функциональности, то и Operations
+2. Тесты можно реализовать с помощью Flapdoodle Embedded MongoDB
+3. Hibernate, равно, как и JPA, и spring-boot-starter-data-jpa не должно остаться в зависимостях, если ДЗ выполняется на основе предыдущего.
+4. Как хранить книги, авторов, жанры и комментарии решать Вам. Но перенесённая с реляционной базы структура не всегда будет подходить для MongoDB.
+
+### Reference Documentation
+For further reference, please consider the following sections:
+
+* [Official Apache Maven documentation](https://maven.apache.org/guides/index.html)
+* [Spring Boot Maven Plugin Reference Guide](https://docs.spring.io/spring-boot/docs/2.2.4.RELEASE/maven-plugin/)
+* [Spring Data JPA](https://docs.spring.io/spring-boot/docs/2.2.4.RELEASE/reference/htmlsingle/#boot-features-jpa-and-spring-data)
+* [Spring Boot DevTools](https://docs.spring.io/spring-boot/docs/2.2.4.RELEASE/reference/htmlsingle/#using-boot-devtools)
+
+### Guides
+The following guides illustrate how to use some features concretely:
+
+* [Accessing Data with JPA](https://spring.io/guides/gs/accessing-data-jpa/)
+
